+++ GenericPM-Spores/include/cropinterfaceS.h
@@ -10,10 +10,10 @@ protected:
     static CropInterfaceS *instance;
     int lastOrgan = 0;
     int plantingDate = 0;
-    double soilMoisture = 0;
-    double newDailySenescenceArea = 0;
+    float soilMoisture = 0;
+    float newDailySenescenceArea = 0;
     bool newOrgan = false;
-    std::vector<std::vector<double>> data;
+    std::vector<std::vector<float>> data;
 
 public:
     static CropInterfaceS* getInstance();
@@ -38,31 +38,31 @@ public:
         return data.size();
     }
 
-    double getOrganArea(int organ) {
+    float getOrganArea(int organ) {
         return data[organ - 1][0];
     }
 
-    double getSenescenceOrganArea(int organ) {
+    float getSenescenceOrganArea(int organ) {
         return data[organ - 1][1];
     }
 
-    double getPreviewsOrganArea(int organ) {
+    float getPreviewsOrganArea(int organ) {
         return data[organ - 1][2];
     }
 
-    double getPreviewsSenescenceOrganArea(int organ) {
+    float getPreviewsSenescenceOrganArea(int organ) {
         return data[organ - 1][3];
     }
 
-    double getRatioDueDefoliation(int organ) const {
+    float getRatioDueDefoliation(int organ) const {
         return data[organ - 1][4];
         ;
     }
 
-    void setOrganAreaS(int organ, double area) {
+    void setOrganAreaS(int organ, float area) {
         //std::cout<<"organ "<<organ<<" area "<<area<<std::endl;
         if (organ > data.size()) {
-            std::vector<double> vetLine;
+            std::vector<float> vetLine;
             vetLine.push_back(area); // OrganS Area
             vetLine.push_back(0); // Senescence Area
             vetLine.push_back(area); // Previews OrganS Area
@@ -76,9 +76,9 @@ public:
         }
     }
 
-    void setSenescenceOrganArea(int organ, double area) {
+    void setSenescenceOrganArea(int organ, float area) {
         if (organ > data.size()) {
-            std::vector<double> vetLine;
+            std::vector<float> vetLine;
             vetLine.push_back(0); // OrganS Area
             vetLine.push_back(area); // Senescence Area
             vetLine.push_back(0); // Previews OrganS Area
@@ -95,9 +95,9 @@ public:
     /* Set Senescence area (cm2/m2) in a given day (DSSAT - PEST - SLDOT)
        This area will be used to remove organs as daily senescence (older first)
     */
-    void setDailySenescenceArea(double area) {
+    void setDailySenescenceArea(float area) {
         newDailySenescenceArea = area;
-        double diff=0;
+        float diff=0;
         for (unsigned int i = 0; i < data.size(); i++) {
             // set the senescence area to each organ and just call organ rate if it has area to be affected 
             //printf("A. OrganS: %d TotalArea: %f SenescenceArea: %f DailySenesc: %f\n",i,data[i][0],data[i][1],newDailySenescenceArea);
@@ -131,13 +131,13 @@ public:
     /* Return the senescence area (cm2/m2) set (DSSAT - PEST - SLDOT)
        This represents the area that need to be removed (organs - older first)
     */
-    double getDailySenescenceArea() {
+    float getDailySenescenceArea() {
         return(newDailySenescenceArea);
     }
 
-    void setRatioDueDefoliation(int organ, double ratio) {
+    void setRatioDueDefoliation(int organ, float ratio) {
         if (organ > data.size()) {
-            std::vector<double> vetLine;
+            std::vector<float> vetLine;
             vetLine.push_back(0); // OrganS Area
             vetLine.push_back(0); // Senescence Area
             vetLine.push_back(0); // Previews OrganS Area
@@ -158,11 +158,11 @@ public:
         return plantingDate;
     }
 
-    void setSoilMoisture(double soilMoisture) {
+    void setSoilMoisture(float soilMoisture) {
         this->soilMoisture = soilMoisture;
     }
     
-    double getSoilMoisture() const {
+    float getSoilMoisture() const {
         return soilMoisture;
     }
